+++ Cargo.toml
@@ -5,6 +5,16 @@ edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
+[lib]
+path = "src/lib.rs"
+
+[[bin]]
+path = "src/main.rs"
+name = "zero2prod"
+
 [dependencies]
 axum = "0.6.15"
 tokio = { version = "1.27.0", features = ["full"] }
+
+[dev-dependencies]
+tower = "0.4.13"
