+++ tests/health_check.rs
@@ -1,0 +1,22 @@
+use axum::body::Body;
+use axum::http::Request;
+use axum::http::StatusCode;
+use tower::ServiceExt;
+
+// #[cfg(feature = "integration_test")]
+#[tokio::test]
+async fn health_check_works() {
+    let app = zero2prod::create_app();
+
+    let response = app
+        .oneshot(
+            Request::builder()
+                .uri("/health_check")
+                .body(Body::empty())
+                .unwrap(),
+        )
+        .await
+        .unwrap();
+
+    assert_eq!(response.status(), StatusCode::OK);
+}
